item service tests without saving

+++ src/main/java/com/example/shopBackend/item/ItemService.java
@@ -1,13 +1,13 @@
 package com.example.shopBackend.item;
 
-import java.util.List;
-
+import com.example.shopBackend.user.UserRepository;
+import exception.BadRequestException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import com.example.shopBackend.review.Review;
+import java.util.List;
 
 /**
  * Services for the item table.
@@ -19,17 +19,23 @@ public class ItemService {
 
 	@Autowired
 	private ItemRepository itemRepository;
-	
+	@Autowired
+	private UserRepository userRepository;
+
+	public ItemService(ItemRepository itemRepository, UserRepository userRepository) {
+		this.itemRepository = itemRepository;
+		this.userRepository = userRepository;
+	}
+
 	/**
 	 * Saves a new item to the database.
 	 * @param {Item} item
 	 * 		  The item to be added to the database.
 	 * @return
 	 */
-	public List<Item> saveReview(List<Item> item) {
+	public List<Item> saveItem(List<Item> item) {
 		return itemRepository.saveAll(item);
 	}
-	
 	/**
 	 * Finds all items for user. And returns them.
 	 * @param {int} id
@@ -39,6 +45,11 @@ public class ItemService {
 	 * @return reviews that match query.
 	 */
 	public List<Item> getItemsForUser(int id, int page) {
+		if(userRepository.findById(id).isEmpty()) {
+			throw new BadRequestException(
+					"No users exists with id " + id);
+		}
+
 		Pageable pageRequest = PageRequest.of(page, 6);
 		return itemRepository.findAllUserId(id, pageRequest);
 	}
@@ -49,7 +60,12 @@ public class ItemService {
 	 * 		  The item to be deleted from the database.
 	 * @return
 	 */
-	public Boolean deleteReview(int id){
+	public Boolean deleteItem(int id){
+		if(itemRepository.findById(id).isEmpty()) {
+			throw new BadRequestException(
+					"No items exists with id " + id);
+		}
+
 		itemRepository.deleteById(id);
 		// check if fails
 		return true;
